dev: remove package tools as clean up
