--- pkg/tools/tools.go
@@ -1,176 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package tools
-
-import (
-	"crypto/ecdsa"
-	"crypto/x509"
-	"encoding/pem"
-	"errors"
-	"fmt"
-	"os"
-
-	"github.com/hyperledger/fabric-admin-sdk/internal/protoutil"
-	"github.com/hyperledger/fabric-admin-sdk/internal/util"
-
-	"github.com/hyperledger/fabric-admin-sdk/internal/configtxgen/encoder"
-	"github.com/hyperledger/fabric-admin-sdk/internal/configtxgen/genesisconfig"
-	"github.com/hyperledger/fabric-admin-sdk/pkg/identity"
-
-	cb "github.com/hyperledger/fabric-protos-go-apiv2/common"
-)
-
-const (
-	msgVersion = int32(0)
-	epoch      = 0
-)
-
-// ConfigTxGen based on Profile return block
-func ConfigTxGen(config *genesisconfig.Profile, channelID string) (*cb.Block, error) {
-	pgen, err := encoder.NewBootstrapper(config)
-	if err != nil {
-		return nil, err
-	}
-	genesisBlock := pgen.GenesisBlockForChannel(channelID)
-	return genesisBlock, nil
-}
-
-// LoadProfile file as file path
-func LoadProfile(configName, fabricConfigPath string) (*genesisconfig.Profile, error) {
-	return genesisconfig.Load(configName, fabricConfigPath)
-}
-
-func CreateSigner(privateKeyPath, certificatePath, mspID string) (identity.SigningIdentity, error) {
-	priv, err := GetPrivateKey(privateKeyPath)
-	if err != nil {
-		return nil, err
-	}
-
-	cert, _, err := GetCertificate(certificatePath)
-	if err != nil {
-		return nil, err
-	}
-
-	return identity.NewPrivateKeySigningIdentity(mspID, cert, priv)
-}
-
-func GetPrivateKey(f string) (*ecdsa.PrivateKey, error) {
-	in, err := os.ReadFile(f)
-	if err != nil {
-		return nil, err
-	}
-
-	k, err := PEMtoPrivateKey(in)
-	if err != nil {
-		return nil, err
-	}
-
-	key, ok := k.(*ecdsa.PrivateKey)
-	if !ok {
-		return nil, errors.New("expecting ecdsa key")
-	}
-
-	return key, nil
-}
-
-func GetCertificate(f string) (*x509.Certificate, []byte, error) {
-	in, err := os.ReadFile(f)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	block, _ := pem.Decode(in)
-
-	c, err := x509.ParseCertificate(block.Bytes)
-	return c, in, err
-}
-
-// PEMtoPrivateKey unmarshals a pem to private key
-func PEMtoPrivateKey(raw []byte) (interface{}, error) {
-	if len(raw) == 0 {
-		return nil, errors.New("invalid PEM. It must be different from nil")
-	}
-	block, _ := pem.Decode(raw)
-	if block == nil {
-		return nil, fmt.Errorf("failed decoding PEM. Block must be different from nil. [% x]", raw)
-	}
-	cert, err := DERToPrivateKey(block.Bytes)
-	if err != nil {
-		return nil, err
-	}
-	return cert, err
-}
-
-// DERToPrivateKey unmarshals a der to private key
-func DERToPrivateKey(der []byte) (key interface{}, err error) {
-
-	if key, err = x509.ParsePKCS1PrivateKey(der); err == nil {
-		return key, nil
-	}
-
-	if key, err = x509.ParsePKCS8PrivateKey(der); err == nil {
-		switch key.(type) {
-		case *ecdsa.PrivateKey:
-			return
-		default:
-			return nil, errors.New("found unknown private key type in PKCS#8 wrapping")
-		}
-	}
-
-	if key, err = x509.ParseECPrivateKey(der); err == nil {
-		return
-	}
-
-	return nil, errors.New("invalid key type. The DER must contain an ecdsa.PrivateKey")
-}
-
-func SignConfigTx(channelID string, envConfigUpdate *cb.Envelope, signer identity.SigningIdentity) (*cb.Envelope, error) {
-	payload, err := protoutil.UnmarshalPayload(envConfigUpdate.Payload)
-	if err != nil {
-		return nil, errors.New("bad payload")
-	}
-
-	if payload.Header == nil || payload.Header.ChannelHeader == nil {
-		return nil, errors.New("bad header")
-	}
-
-	ch, err := protoutil.UnmarshalChannelHeader(payload.Header.ChannelHeader)
-	if err != nil {
-		return nil, errors.New("could not unmarshall channel header")
-	}
-
-	if ch.Type != int32(cb.HeaderType_CONFIG_UPDATE) {
-		return nil, errors.New("bad type")
-	}
-
-	if ch.ChannelId == "" {
-		return nil, errors.New("empty channel id")
-	}
-
-	configUpdateEnv, err := protoutil.UnmarshalConfigUpdateEnvelope(payload.Data)
-	if err != nil {
-		return nil, errors.New("bad config update env")
-	}
-
-	sigHeader, err := protoutil.NewSignatureHeader(signer)
-	if err != nil {
-		return nil, err
-	}
-
-	configSig := &cb.ConfigSignature{
-		SignatureHeader: protoutil.MarshalOrPanic(sigHeader),
-	}
-
-	configSig.Signature, err = signer.Sign(util.Concatenate(configSig.SignatureHeader, configUpdateEnv.ConfigUpdate))
-	if err != nil {
-		return nil, err
-	}
-
-	configUpdateEnv.Signatures = append(configUpdateEnv.Signatures, configSig)
-
-	return protoutil.CreateSignedEnvelope(cb.HeaderType_CONFIG_UPDATE, channelID, signer, configUpdateEnv, msgVersion, epoch)
-}
