--- pkg/tools/tools_suite_test.go
@@ -1,13 +1,0 @@
-package tools_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestTools(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Tools Suite")
-}
